Move printKeyValue
